% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partial_Moments.R
\name{Co_UPM}
\alias{Co_UPM}
\title{Co-Upper Partial Moment
(Upper Right Quadrant 1)}
\usage{
Co_UPM(degree, target1, target2, variable1, variable2)
}
\arguments{
\item{degree}{Degree = 0 is frequency, degree = 1 is area}

\item{target1}{Typically set to mean of Variable 1, but does not have to be}

\item{target2}{Typically set to mean of Variable 2, but does not have to be}

\item{variable1}{Variable 1}

\item{variable2}{Variable 2}
}
\value{
Co-UPM of two variables
}
\description{
This function generates a multivariate co-upper partial moment for any degree or target.
}
\examples{
set.seed(123)
x<-rnorm(100); y<-rnorm(100)
\dontrun{Co_UPM(0,mean(x),mean(y),x,y)}
}
\keyword{moments}
\keyword{partial}

