% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{D.LPM}
\alias{D.LPM}
\title{Divergent-Lower Partial Moment
(Lower Right Quadrant 3)}
\usage{
D.LPM(degree_lpm, degree_upm, x, y, target_x, target_y)
}
\arguments{
\item{degree_lpm}{numeric; Degree for lower deviations of variable Y.  \code{(degree_lpm = 0)} is frequency, \code{(degree_lpm = 1)} is area.}

\item{degree_upm}{numeric; Degree for upper deviations of variable X.  \code{(degree_upm = 0)} is frequency, \code{(degree_upm = 1)} is area.}

\item{x}{a numeric vector.   \link{data.frame} or \link{list} type objects are not permissible.}

\item{y}{a numeric vector of equal length to \code{x}.   \link{data.frame} or \link{list} type objects are not permissible.}

\item{target_x}{numeric; Target for upside deviations of variable X.  Typically the mean of Variable X for classical statistics equivalences, but does not have to be.}

\item{target_y}{numeric; Target for lower deviations of variable Y.  Typically the mean of Variable Y for classical statistics equivalences, but does not have to be.}
}
\value{
Divergent LPM of two variables
}
\description{
This function generates a divergent lower partial moment between two equal length variables for any degree or target.
}
\examples{
set.seed(123)
x <- rnorm(100) ; y <- rnorm(100)
D.LPM(0, 0, x, y, mean(x), mean(y))
}
\references{
Viole, F. and Nawrocki, D. (2013) "Nonlinear Nonparametric Statistics: Using Partial Moments" (ISBN: 1490523995)
}
\author{
Fred Viole, OVVO Financial Systems
}
