% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_probability.R
\name{Feature.probability}
\alias{Feature.probability}
\title{Feature Probability}
\usage{
Feature.probability(x, y, threshold = 0, point.est = NULL)
}
\arguments{
\item{x}{Complete cleaned dataset in matrix form.}

\item{y}{Column of data to be classified.}

\item{threshold}{Sets the correlation threshold for independent variables.  Defaults to 0.}

\item{point.est}{IV data point(s) to be classified, in matrix form.}
}
\value{
Returns two variables, mean squared error "\code{MSE}" and predicted values "\code{Predictions}" as well as Prediction Accuracy measured by percentage of exact classifications.
}
\description{
Classifies data based on feature probabilities
}
\examples{
## Using 'iris' dataset where predictive attributes are columns 1:4, and the class is column 5.
Feature.probability(iris,5)

## To call mean squared error
Feature.probability(iris,5)$MSE

## To call fitted values
Feature.probability(iris,5)$Fitted

## To generate a single predicted value
Feature.probability(iris,5, point.est=cbind(5.1,3.5,1.4,0.2))$prediction

## To generate multiple predicted values
Feature.probability(iris,5, point.est=(iris[1:10,1:4]))$prediction
}
\author{
Fred Viole, OVVO Financial Systems
}
\references{
Annotated code is available at
\url{https://github.com/OVVO-Financial/NNS/blob/Prelim/R/Feature_probability.R}
}
\keyword{classifier}

