% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find_Interval.R
\name{findInterval2}
\alias{findInterval2}
\title{Find Interval}
\usage{
findInterval2(x, v)
}
\arguments{
\item{x}{Variable / Vector}

\item{v}{Vector}
}
\value{
Returns interval number
}
\description{
Internal Function Only.
Removes the sort check from the base function \link{findInterval}
}

