% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uni_SD_Routines.R
\name{FSD}
\alias{FSD}
\title{FSD}
\usage{
FSD(x, y)
}
\arguments{
\item{x}{variable}

\item{y}{variable}
}
\description{
Uni-directional test of first degree stochastic dominance using lower partial moments used in SD Efficient Set routine.
}
\examples{
set.seed(123)
x<-rnorm(100); y<-rnorm(100)
\dontrun{FSD(x,y)}
}
\author{
Fred Viole, OVVO Financial Systems
}
\references{
Viole, F. and Nawrocki, D. (2016) "LPM Density Functions for the Computation of the SD Efficient Set." Journal of Mathematical Finance, 6, 105-126. \url{http://www.scirp.org/Journal/PaperInformation.aspx?PaperID=63817}.
}

