% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPM_UPM_VaR.R
\name{LPM.VaR}
\alias{LPM.VaR}
\title{LPM VaR}
\usage{
LPM.VaR(percentile, degree, x)
}
\arguments{
\item{percentile}{numeric [0, 1]; The percentile for left-tail VaR.}

\item{degree}{integer; \code{(degree = 0)} for discrete distributions, \code{(degree = 1)} for continuous distributions.}

\item{x}{a numeric vector.}
}
\value{
Returns a numeric value representing the point at which \code{"percentile"} of the area of \code{x} is below.
}
\description{
Generates a value at risk (VaR) quantile based on the Lower Partial Moment ratio.
}
\examples{
\dontrun{
set.seed(123)
x <- rnorm(100)

## For 5th percentile, left-tail
LPM.VaR(0.05, 0, x)
}
}
\references{
Viole, F. and Nawrocki, D. (2013) "Nonlinear Nonparametric Statistics: Using Partial Moments" (ISBN: 1490523995, 2nd edition: \url{https://ovvo-financial.github.io/NNS/book/})
}
\author{
Fred Viole, OVVO Financial Systems
}
