% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binary_ANOVA.R
\name{NNS.ANOVA.bin}
\alias{NNS.ANOVA.bin}
\title{NNS ANOVA Binary}
\usage{
NNS.ANOVA.bin(control, treatment, confidence.interval = NULL)
}
\arguments{
\item{control}{The control group sample}

\item{treatment}{The treatment group sample}

\item{confidence.interval}{The confidence interval surrounding the control mean.  Defaults to \code{confidence.interval=NULL}.}
}
\value{
Returns \code{"Control Mean"}, \code{"Treatment Mean"}, \code{"Grand Mean"}, \code{"Control CDF"}, \code{"Treatment CDF"}, the certainty of the same population statistic \code{"Certainty"}, the effect size of the treatment for a specified confidence interval with \code{"Lower Bound Effect"} and \code{"Upper Bound Effect"}.
}
\description{
Performs an analysis of variance (ANOVA) for two variables: control and treatment.  Returns the effect size of the treatment for a specified confidence interval.
}
\examples{
set.seed(123)
x<-rnorm(100); y<-rnorm(100)
NNS.ANOVA.bin(x,y,0.95)
}
\keyword{ANOVA,}
\keyword{effect}
\keyword{size}

