% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dependence_base.R
\name{NNS.dep.base}
\alias{NNS.dep.base}
\title{NNS Dependence Base}
\usage{
NNS.dep.base(
  x,
  y = NULL,
  order = NULL,
  degree = NULL,
  type = NULL,
  print.map = FALSE,
  asym = FALSE
)
}
\arguments{
\item{x}{from \link{NNS.part}}

\item{y}{from \link{NNS.part}}

\item{order}{from \link{NNS.part}}

\item{degree}{from \link{NNS.part}}

\item{type}{from \link{NNS.part}}

\item{print.map}{from \link{NNS.part}}

\item{asym}{for asymmetrical dependecies}
}
\value{
Returns NNS dependence.
}
\description{
Internal function for NNS dependence \link{NNS.dep} parallel instances.
}
