% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dependence_matrix.R
\name{NNS.dep.matrix}
\alias{NNS.dep.matrix}
\title{NNS Dependence Matrix (INTERNAL CALL FOR \link{NNS.dep})}
\usage{
NNS.dep.matrix(x, order = NULL, degree = NULL)
}
\arguments{
\item{x}{Variable 1 matrix}

\item{order}{Controls the level of quadrant partitioning.  Default to \code{order=NULL}.  Errors can generally be rectified by setting \code{order=1}.}

\item{degree}{Defaults to NULL to allow number of observations to be \code{degree} determinant.}
}
\value{
Returns the \code{"Correlation"} and \code{"Dependence"}
}
\description{
Returns the dependence matrix for \link{NNS.dep}.
}
\author{
Fred Viole, OVVO Financial Systems
}
\keyword{correlation}
\keyword{dependence,}

