% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNS_Distance_bulk.R
\name{NNS.distance.path.bulk}
\alias{NNS.distance.path.bulk}
\title{All-k bulk predictor (parallel if ncores > 1)}
\usage{
NNS.distance.path.bulk(rpm, Xtest, kmax, class = NULL, ncores = 1L)
}
\description{
All-k bulk predictor (parallel if ncores > 1)
}
