% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uni_SD_Routines.R
\name{NNS.FSD.uni}
\alias{NNS.FSD.uni}
\title{NNS FSD Test uni-directional}
\usage{
NNS.FSD.uni(x, y, type = "discrete")
}
\arguments{
\item{x}{a numeric vector.}

\item{y}{a numeric vector.}

\item{type}{options: ("discrete", "continuous"); \code{"discrete"} (default) selects the type of CDF.}
}
\value{
Returns (1) if \code{"X FSD Y"}, else (0).
}
\description{
Uni-directional test of first degree stochastic dominance using lower partial moments used in SD Efficient Set routine.
}
\examples{
set.seed(123)
x <- rnorm(100) ; y <- rnorm(100)
NNS.FSD.uni(x, y)
}
\references{
Viole, F. and Nawrocki, D. (2016) "LPM Density Functions for the Computation of the SD Efficient Set." Journal of Mathematical Finance, 6, 105-126. \url{http://www.scirp.org/Journal/PaperInformation.aspx?PaperID=63817}.

Viole, F. (2017) "A Note on Stochastic Dominance." \url{https://ssrn.com/abstract=3002675}.
}
\author{
Fred Viole, OVVO Financial Systems
}
