% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Central_tendencies.R
\name{NNS.gravity}
\alias{NNS.gravity}
\title{NNS gravity}
\usage{
NNS.gravity(x, discrete = FALSE)
}
\arguments{
\item{x}{vector of data.}

\item{discrete}{logical; \code{FALSE} (default) for discrete distributions.}
}
\value{
Returns a numeric value representing the central tendency of the distribution.
}
\description{
Alternative central tendency measure more robust to outliers.
}
\examples{
\dontrun{
set.seed(123)
x <- rnorm(100)
NNS.gravity(x)
}
}
\author{
Fred Viole, OVVO Financial Systems
}
