% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNS-package.R
\docType{package}
\name{NNS}
\alias{NNS}
\alias{_PACKAGE}
\alias{NNS-package}
\title{NNS: Nonlinear Nonparametric Statistics}
\description{
Nonlinear nonparametric statistics using partial moments.  Partial moments are the elements of variance and asymptotically approximate the area of f(x).  These robust statistics provide the basis for nonlinear analysis while retaining linear equivalences.  NNS offers: Numerical integration, Numerical differentiation, Clustering, Correlation, Dependence, Causal analysis, ANOVA, Regression, Classification, Seasonality, Autoregressive modeling, Normalization and Stochastic dominance.  All routines based on: Viole, F. and Nawrocki, D. (2013), Nonlinear Nonparametric Statistics: Using Partial Moments (ISBN: 1490523995, 2nd edition: \url{https://ovvo-financial.github.io/NNS/book/}).
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/OVVO-Financial/NNS/issues}
}

}
\author{
\strong{Maintainer}: Fred Viole \email{ovvo.open.source@gmail.com}

Other contributors:
\itemize{
  \item Roberto Spadim [contributor]
}

}
\keyword{internal}
