% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSD.R
\name{NNS.SSD}
\alias{NNS.SSD}
\title{NNS SSD Test}
\usage{
NNS.SSD(x, y, plot = TRUE)
}
\arguments{
\item{x}{a numeric vector.}

\item{y}{a numeric vector.}

\item{plot}{logical; \code{TRUE} (default) plots the SSD test.}
}
\value{
Returns one of the following SSD results: \code{"X SSD Y"}, \code{"Y SSD X"}, or \code{"NO SSD EXISTS"}.
}
\description{
Bi-directional test of second degree stochastic dominance using lower partial moments.
}
\examples{
\dontrun{
set.seed(123)
x <- rnorm(100) ; y <- rnorm(100)
NNS.SSD(x, y)
}
}
\references{
Viole, F. and Nawrocki, D. (2016) "LPM Density Functions for the Computation of the SD Efficient Set." Journal of Mathematical Finance, 6, 105-126.  \doi{10.4236/jmf.2016.61012}.
}
\author{
Fred Viole, OVVO Financial Systems
}
