% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uni_SD_Routines.R
\name{NNS.SSD.uni}
\alias{NNS.SSD.uni}
\title{NNS SSD Test uni-directional}
\usage{
NNS.SSD.uni(x, y)
}
\arguments{
\item{x}{a numeric vector.}

\item{y}{a numeric vector.}
}
\value{
Returns (1) if \code{"X SSD Y"}, else (0).
}
\description{
Uni-directional test of second degree stochastic dominance using lower partial moments used in SD Efficient Set routine.
}
\examples{
set.seed(123)
x <- rnorm(100) ; y <- rnorm(100)
NNS.SSD.uni(x, y)
}
\references{
Viole, F. and Nawrocki, D. (2016) "LPM Density Functions for the Computation of the SD Efficient Set." Journal of Mathematical Finance, 6, 105-126. \url{http://www.scirp.org/Journal/PaperInformation.aspx?PaperID=63817}.
}
\author{
Fred Viole, OVVO Financial Systems
}
