% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uni_SD_Routines.R
\name{NNS.TSD.uni}
\alias{NNS.TSD.uni}
\title{NNS TSD Test uni-directional}
\usage{
NNS.TSD.uni(x, y)
}
\arguments{
\item{x}{a numeric vector.}

\item{y}{a numeric vector.}
}
\value{
Returns (1) if \code{"X TSD Y"}, else (0).
}
\description{
Uni-directional test of third degree stochastic dominance using lower partial moments used in SD Efficient Set routine.
}
\examples{
set.seed(123)
x<-rnorm(100); y<-rnorm(100)
NNS.TSD.uni(x,y)
}
\author{
Fred Viole, OVVO Financial Systems
}
