% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uni_SD_Routines.R
\name{SSD}
\alias{SSD}
\title{SSD}
\usage{
SSD(x, y)
}
\arguments{
\item{x}{variable}

\item{y}{variable}
}
\description{
Uni-directional test of second degree stochastic dominance using lower partial moments used in SD Efficient Set routine.
}
\examples{
set.seed(123)
x<-rnorm(100); y<-rnorm(100)
\dontrun{SSD(x,y)}
}
\author{
Fred Viole, OVVO Financial Systems
}

