% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uni_SD_Routines.R
\name{TSD}
\alias{TSD}
\title{TSD}
\usage{
TSD(x, y)
}
\arguments{
\item{x}{variable}

\item{y}{variable}
}
\description{
Uni-directional test of third degree stochastic dominance using lower partial moments used in SD Efficient Set routine.
}
\examples{
set.seed(123)
x<-rnorm(100); y<-rnorm(100)
\dontrun{TSD(x,y)}
}
\author{
Fred Viole, OVVO Financial Systems
}

