% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uni_Causation.R
\name{Uni.caus}
\alias{Uni.caus}
\title{NNS Causation uni-directional  (INTERNAL CALL FOR \link{NNS.caus})}
\usage{
Uni.caus(x, y, tau)
}
\arguments{
\item{x}{Variable}

\item{y}{Variable}

\item{tau}{Number of lagged observations to consider}
}
\description{
Returns the uni-directional causality from observational data between two variables.  Causation nets out the univariate effect.
}
\author{
Fred Viole, OVVO Financial Systems
}
\keyword{causation}

