% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binary_ANOVA.R
\name{VN.ANOVA.bin}
\alias{VN.ANOVA.bin}
\title{VN ANOVA Binary}
\usage{
VN.ANOVA.bin(control, treatment, confidence.interval)
}
\arguments{
\item{control}{The control group sample}

\item{treatment}{The treatment group sample}

\item{confidence.interval}{The confidence interval surrounding the control mean.}
}
\description{
Performs an analysis of variance (ANOVA) for two variables: control and treatment.  Returns the effect size of the treatment for a specified confidence interval.
}
\examples{
set.seed(123)
x<-rnorm(100); y<-rnorm(100)
VN.ANOVA.bin(x,y,0.95)
}

