% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA.R
\name{VN.ANOVA}
\alias{VN.ANOVA}
\title{VN ANOVA}
\usage{
VN.ANOVA(A)
}
\arguments{
\item{A}{Matrix of variables.}
}
\description{
Analysis of variance (ANOVA) based on lower partial moment CDFs for multiple variables.  Returns a degree of certainty the samples belong to the same population, not a p-value.
}
\examples{
set.seed(123)
x<-rnorm(100); y<-rnorm(100)
A<-cbind(x,y)
\dontrun{VN.ANOVA(A)}
}

