% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Causation.R
\name{VN.Causation}
\alias{VN.Causation}
\title{VN Causation}
\usage{
VN.Causation(x, y, tau)
}
\arguments{
\item{x}{Variable}

\item{y}{Variable}

\item{tau}{Number of lagged observations to consider}
}
\description{
Returns the causality from observational data between two variables
}
\examples{
set.seed(123)
x<-rnorm(100); y<-rnorm(100)
\dontrun{VN.Causation(x,y,3)}
}
\keyword{causation}

