% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSD.R
\name{VN.FSD}
\alias{VN.FSD}
\title{VN FSD Test}
\usage{
VN.FSD(x, y)
}
\arguments{
\item{x}{variable}

\item{y}{variable}
}
\description{
Bi-directional test of first degree stochastic dominance using lower partial moments.
}
\examples{
set.seed(123)
x<-rnorm(100); y<-rnorm(100)
VN.FSD(x,y)
}
\author{
Fred Viole, OVVO Financial Systems
}
\references{
Viole, F. and Nawrocki, D. (2016) "LPM Density Functions for the Computation of the SD Efficient Set." Journal of Mathematical Finance, 6, 105-126. \url{http://www.scirp.org/Journal/PaperInformation.aspx?PaperID=63817}.
}

