% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization_Causation.R
\name{VN.norm.caus}
\alias{VN.norm.caus}
\title{VN.norm.caus}
\usage{
VN.norm.caus(A)
}
\arguments{
\item{A}{Matrix of variables.}
}
\description{
VN.norm function with reduced output specifically for VN.caus routine
}
\examples{
set.seed(123)
x<-rnorm(100); y<-rnorm(100)
A<-cbind(x,y)
\dontrun{VN.norm.caus(A)}
}
\author{
Fred Viole, OVVO Financial Systems
}
\references{
Viole, F. and Nawrocki, D. (2013) "Nonlinear Nonparametric Statistics: Using Partial Moments"
\url{http://amzn.com/1490523995}
}

