% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seasonality_Test.R
\name{VN.seas}
\alias{VN.seas}
\title{VN Seasonality Test}
\usage{
VN.seas(variable)
}
\arguments{
\item{variable}{Variable}
}
\description{
Seasonality test based on the coefficient of variance for the variable and lagged component series.  A result of 1 signifies no seasonality present.
}
\examples{
set.seed(123)
x<-rnorm(100)
VN.seas(x)
}
\author{
Fred Viole, OVVO Financial Systems
}
\references{
Viole, F. and Nawrocki, D. (2013) "Nonlinear Nonparametric Statistics: Using Partial Moments"
\url{http://amzn.com/1490523995}
}

