% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partial_Moments.R
\name{Co.LPM}
\alias{Co.LPM}
\title{Co-Lower Partial Moment
(Lower Left Quadrant 4)}
\usage{
Co.LPM(degree, variable1, variable2, target1 = mean(variable1),
  target2 = mean(variable2))
}
\arguments{
\item{degree}{Degree = 0 is frequency, degree = 1 is area}

\item{variable1}{Variable 1}

\item{variable2}{Variable 2}

\item{target1}{Defaults to mean of Variable 1, but does not have to be...}

\item{target2}{Defualts to mean of Variable 2, but does not have to be...}
}
\value{
Co-LPM of two variables
}
\description{
This function generates a multivariate co-lower partial moment for any degree or target.
}
\examples{
set.seed(123)
x<-rnorm(100); y<-rnorm(100)
Co.LPM(0,mean(x),mean(y),x,y)
}
\author{
Fred Viole, OVVO Financial Systems
}
\references{
Viole, F. and Nawrocki, D. (2013) "Nonlinear Nonparametric Statistics: Using Partial Moments"
\url{http://amzn.com/1490523995}
}
\keyword{covariance}
\keyword{moments,}
\keyword{partial}

