% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partial_Moments.R
\name{Co.LPM_nD}
\alias{Co.LPM_nD}
\title{Co‑Lower Partial Moment nD}
\usage{
Co.LPM_nD(data, target, degree = 0)
}
\arguments{
\item{data}{A numeric matrix with observations in rows and variables in columns.}

\item{target}{A numeric vector of length equal to the number of columns in `data`.}

\item{degree}{Numeric; degree for lower deviations (0 = frequency, 1 = area).}
}
\value{
Numeric; the n‑dimensional co‑lower partial moment, normalized.
}
\description{
This function generates a co‑lower partial moment for n‑dimensional data (n >= 2) for any degree or target.
}
\examples{
mat <- matrix(rnorm(200), ncol = 4)
Co.LPM_nD(mat, rep(0, ncol(mat)), 1)
}
