% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{D.UPM}
\alias{D.UPM}
\title{Divergent-Upper Partial Moment
(Upper Left Quadrant 2)}
\usage{
D.UPM(degree_x, degree_y, x, y, target_x, target_y)
}
\arguments{
\item{degree_x}{integer; Degree for variable X.  \code{(degree_x = 0)} is frequency, \code{(degree_x = 1)} is area.}

\item{degree_y}{integer; Degree for variable Y.  \code{(degree_y = 0)} is frequency, \code{(degree_y = 1)} is area.}

\item{x}{a numeric vector.   \link{data.frame} or \link{list} type objects are not permissible.}

\item{y}{a numeric vector of equal length to \code{x}.   \link{data.frame} or \link{list} type objects are not permissible.}

\item{target_x}{numeric; Typically the mean of Variable X for classical statistics equivalences, but does not have to be.}

\item{target_y}{numeric; Typically the mean of Variable Y for classical statistics equivalences, but does not have to be.}
}
\value{
Divergent UPM of two variables
}
\description{
This function generates a divergent upper partial moment between two equal length variables for any degree or target.
}
\examples{
set.seed(123)
x <- rnorm(100) ; y <- rnorm(100)
D.UPM(0, 0, x, y, mean(x), mean(y))
}
\references{
Viole, F. and Nawrocki, D. (2013) "Nonlinear Nonparametric Statistics: Using Partial Moments"
\url{https://www.amazon.com/dp/1490523995/ref=cm_sw_su_dp}
}
\author{
Fred Viole, OVVO Financial Systems
}
