% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partial_Moments.R
\name{LPM}
\alias{LPM}
\title{Lower Partial Moment}
\usage{
LPM(degree, target, variable)
}
\arguments{
\item{degree}{Degree = 0 is frequency, degree = 1 is area}

\item{target}{Typically set to mean, but does not have to be}

\item{variable}{Variable}
}
\value{
LPM of variable
}
\description{
This function generates a univariate lower partial moment for any degree or target.
}
\examples{
set.seed(123)
x<-rnorm(100)
\dontrun{LPM(0,0,x)}
}
\keyword{moments}
\keyword{partial}

