% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partial_Moments.R
\name{LPM}
\alias{LPM}
\title{Lower Partial Moment}
\usage{
LPM(degree, target, variable, excess_ret = FALSE)
}
\arguments{
\item{degree}{integer; \code{(degree = 0)} is frequency, \code{(degree = 1)} is area.}

\item{target}{numeric; Set to \code{target = mean(variable)} for classical equivalences, but does not have to be. (Vectorized)}

\item{variable}{a numeric vector.   \link{data.frame} or \link{list} type objects are not permissible.}

\item{excess_ret}{Logical; \code{FALSE} (default)}
}
\value{
LPM of variable
}
\description{
This function generates a univariate lower partial moment for any degree or target.
}
\examples{
set.seed(123)
x <- rnorm(100)
LPM(0, mean(x), x)
}
\references{
Viole, F. and Nawrocki, D. (2013) "Nonlinear Nonparametric Statistics: Using Partial Moments" (ISBN: 1490523995)
}
\author{
Fred Viole, OVVO Financial Systems
}
