% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partial_Moments.R
\name{LPM}
\alias{LPM}
\title{Lower Partial Moment}
\usage{
LPM(degree, target, variable)
}
\arguments{
\item{degree}{Degree = 0 is frequency, degree = 1 is area}

\item{target}{Typically set to mean, but does not have to be}

\item{variable}{Variable}
}
\value{
LPM of variable
}
\description{
This function generates a univariate lower partial moment for any degree or target.
}
\examples{
set.seed(123)
x<-rnorm(100)
LPM(0,mean(x),x)
}
\author{
Fred Viole, OVVO Financial Systems
}
\references{
Viole, F. and Nawrocki, D. (2013) "Nonlinear Nonparametric Statistics: Using Partial Moments"
\url{http://amzn.com/1490523995}
}
\keyword{moments}
\keyword{partial}

