% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LPM}
\alias{LPM}
\title{Lower Partial Moment}
\usage{
LPM(degree, target, variable)
}
\arguments{
\item{degree}{integer; \code{(degree = 0)} is frequency, \code{(degree = 1)} is area.}

\item{target}{numeric; Typically set to mean, but does not have to be. (Vectorized)}

\item{variable}{a numeric vector.  \link{data.frame} or \link{list} type objects are not permissible.}
}
\value{
LPM of variable
}
\description{
This function generates a univariate lower partial moment for any degree or target.
}
\examples{
set.seed(123)
x <- rnorm(100)
LPM(0, mean(x), x)
}
\references{
Viole, F. and Nawrocki, D. (2013) "Nonlinear Nonparametric Statistics: Using Partial Moments"
\url{https://www.amazon.com/dp/1490523995/ref=cm_sw_su_dp}
}
\author{
Fred Viole, OVVO Financial Systems
}
