% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMA_optim.R
\name{NNS.ARMA.optim}
\alias{NNS.ARMA.optim}
\title{NNS ARMA Optimizer}
\usage{
NNS.ARMA.optim(variable, training.set, seasonal.factor)
}
\arguments{
\item{variable}{a numeric vector.}

\item{training.set}{numeric; \code{NULL} (defualt) Sets the number of variable observations}

\item{seasonal.factor}{integers; Multiple frequency integers considered for \link{NNS.ARMA} model, i.e. \code{(seasonal.factor=c(12,24,36))}}
}
\value{
Returns a list containing a vector of optimal seasonal periods \code{$period} and the minimum SSE value \code{$SSE}.
}
\description{
Wrapper function for optimizing any combination of given `seasonal.periods` in \link{NNS.ARMA}.  Minimum sum of squared errors (forecast-actual) is used to determine optimum.
}
\note{
The number of combinations will grow prohibitively large, they should be kept to a minimum.

\code{seasonal.factor} containing an element too large will result in an error.  Please reduce the maximum \code{seasonal.factor}.
}
\examples{

## Nonlinear NNS.ARMA using AirPassengers monthly data
\dontrun{
NNS.ARMA.optim(AirPassengers,training.set=132,seasonal.periods=seq(12,60,12))
}

}
\references{
Viole, F. and Nawrocki, D. (2013) "Nonlinear Nonparametric Statistics: Using Partial Moments"
\url{http://amzn.com/1490523995}
}
\author{
Fred Viole, OVVO Financial Systems
}
\keyword{Autoregressive}
\keyword{model}
