% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSD.R
\name{NNS.SSD}
\alias{NNS.SSD}
\title{NNS SSD Test}
\usage{
NNS.SSD(x, y)
}
\arguments{
\item{x}{a numeric vector.}

\item{y}{a numeric vector.}
}
\value{
Returns one of the following SSD results: \code{"X SSD Y"}, \code{"Y SSD X"}, or \code{"NO SSD EXISTS"}.
}
\description{
Bi-directional test of second degree stochastic dominance using lower partial moments.
}
\examples{
set.seed(123)
x<-rnorm(100); y<-rnorm(100)
NNS.SSD(x,y)
}
\author{
Fred Viole, OVVO Financial Systems
}
\references{
Viole, F. and Nawrocki, D. (2016) "LPM Density Functions for the Computation of the SD Efficient Set." Journal of Mathematical Finance, 6, 105-126. \url{http://www.scirp.org/Journal/PaperInformation.aspx?PaperID=63817}.
}
\keyword{dominance}
\keyword{stochastic}

