% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Causation.R
\name{Uni.caus}
\alias{Uni.caus}
\title{Uni-Directional Causation}
\usage{
Uni.caus(x, y, tau)
}
\arguments{
\item{x}{Variable}

\item{y}{Variable}

\item{tau}{Number of lagged observations to consider}
}
\description{
Returns the uni-directional causality from observational data between two variables.  Causation nets out the univariate effect.
}
\examples{
set.seed(123)
x<-rnorm(100); y<-rnorm(100)
Uni.caus(x,y,3)
}
\author{
Fred Viole, OVVO Financial Systems
}
\keyword{causation}

