% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binary_ANOVA.R
\name{UPM.VaR}
\alias{UPM.VaR}
\title{UPM VaR}
\usage{
UPM.VaR(percentile, degree, x)
}
\arguments{
\item{percentile}{The percentile for VaR}

\item{degree}{Degree = 0 for discrete distributions, degree = 1 for continuous distributions}

\item{x}{Variable}
}
\description{
Generates an upside VaR based on the Upper Partial Moment ratio
}
\examples{
set.seed(123)
x<-rnorm(100)
\dontrun{UPM.VaR(0.95,0,x)}
}

