% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPM_UPM_VaR.R
\name{UPM.VaR}
\alias{UPM.VaR}
\title{UPM VaR}
\usage{
UPM.VaR(percentile, degree, x)
}
\arguments{
\item{percentile}{numeric [0,1]; The percentile for VaR.}

\item{degree}{integer; \code{(degree=0)} for discrete distributions, \code{(degree=1)} for continuous distributions.}

\item{x}{a numeric vector.}
}
\description{
Generates an upside VaR based on the Upper Partial Moment ratio
}
\examples{
set.seed(123)
x<-rnorm(100)
UPM.VaR(0.95,0,x)
}
\keyword{VaR}

