% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMA_Regression.R
\name{VN.ARMA.reg}
\alias{VN.ARMA.reg}
\title{VN.ARMA.reg}
\usage{
VN.ARMA.reg(x, y, order = NULL, point.est = NULL)
}
\arguments{
\item{x}{Independent Variable}

\item{y}{Dependent Variable}

\item{order}{Number of partial moment quadrants to be generated}

\item{point.est}{Value to be fitted}
}
\description{
VN.reg function with reduced output specifically for VN.ARMA routine
}
\examples{
set.seed(123)
x<-rnorm(100); y<-rnorm(100)
\dontrun{VN.ARMA.reg(x,y)}
}
\author{
Fred Viole, OVVO Financial Systems
}
\references{
Viole, F. and Nawrocki, D. (2013) "Nonlinear Nonparametric Statistics: Using Partial Moments"
\url{http://amzn.com/1490523995}
}

