% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Causation.R
\name{VN.caus}
\alias{VN.caus}
\title{VN Causation}
\usage{
VN.caus(x, y, tau)
}
\arguments{
\item{x}{Variable}

\item{y}{Variable}

\item{tau}{Number of lagged observations to consider}
}
\value{
Returns the directional causation and quantity of association.
}
\description{
Returns the causality from observational data between two variables
}
\examples{
set.seed(123)
x<-rnorm(100); y<-rnorm(100)
\dontrun{VN.caus(x,y,3)}
}
\author{
Fred Viole, OVVO Financial Systems
}
\references{
Viole, F. and Nawrocki, D. (2013) "Nonlinear Nonparametric Statistics: Using Partial Moments"
\url{http://amzn.com/1490523995}
}
\keyword{causation}

