% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{VN.norm}
\alias{VN.norm}
\title{VN Normalization}
\usage{
VN.norm(A, chart.type = NULL)
}
\arguments{
\item{A}{Matrix of variables.}

\item{chart.type}{Defaults to NULL.  'l' for line, 'b' for boxplot}
}
\description{
Normalizes a matrix of variables based on nonlinear scaling normalization method.
}
\examples{
set.seed(123)
x<-rnorm(100); y<-rnorm(100)
A<-cbind(x,y)
VN.norm(A)
}
\author{
Fred Viole, OVVO Financial Systems
}
\references{
Viole, F. and Nawrocki, D. (2013) "Nonlinear Nonparametric Statistics: Using Partial Moments"
\url{http://amzn.com/1490523995}
}

