% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seasonality_Test.R
\name{VN.seas}
\alias{VN.seas}
\title{VN Seasonality Test}
\usage{
VN.seas(variable)
}
\arguments{
\item{variable}{Variable}
}
\description{
Seasonality test based on the coefficient of variance for the variable and lagged component series.  A result of 1 signifies no seasonality present.
}
\examples{
set.seed(123)
x<-rnorm(100)
\dontrun{VN.seas(x)}
}

